调整目录结构

+++ server-code/src/service/zone_service/item/ItemData.cpp
@@ -52,7 +52,7 @@ bool CItemData::Init(CMysqlConnection* pDB, OBJID idItem)
 	CHECKF(m_pRecord != nullptr);
 	std::string jsonTxt = m_pRecord->Field(TBLD_ITEM::EXTRA);
 	pb_util::LoadFromJsonTxt(jsonTxt, m_ExtraData);
-	m_pRecord->Field(TBLD_ITEM::EXTRA).SetGetValStringFunc( std::bind(&SaveItemExtraToJson, std::cref(m_ExtraData)) );
+	m_pRecord->Field(TBLD_ITEM::EXTRA).SetGetValStringFunc(std::bind(&SaveItemExtraToJson, std::cref(m_ExtraData)));
 
 	m_pType = ItemTypeSet()->QueryObj(GetType());
 	if(!m_pType)
@@ -101,7 +101,7 @@ bool CItemData::Init(CMysqlConnection* pDB, ST_ITEMINFO& info)
 
 	if(m_pRecord->Update() == false)
 		return false;
-	m_pRecord->Field(TBLD_ITEM::EXTRA).SetGetValStringFunc( std::bind(&SaveItemExtraToJson, std::cref(m_ExtraData)) );
+	m_pRecord->Field(TBLD_ITEM::EXTRA).SetGetValStringFunc(std::bind(&SaveItemExtraToJson, std::cref(m_ExtraData)));
 	uint32_t nAddition = GetAddition();
 	if(nAddition > 0)
 	{
@@ -131,7 +131,7 @@ bool CItemData::Init(CMysqlConnection* pDB, uint64_t idPlayer, uint32_t idItemTy
 	m_pRecord->Field(TBLD_ITEM::PILENUM)  = nNum;
 	m_pRecord->Field(TBLD_ITEM::POSITION) = nPosition;
 	m_pRecord->Field(TBLD_ITEM::FLAG)	  = dwFlag;
-	m_pRecord->Field(TBLD_ITEM::EXTRA).SetGetValStringFunc( std::bind(&SaveItemExtraToJson, std::cref(m_ExtraData)) );
+	m_pRecord->Field(TBLD_ITEM::EXTRA).SetGetValStringFunc(std::bind(&SaveItemExtraToJson, std::cref(m_ExtraData)));
 	return true;
 	__LEAVE_FUNCTION
 	return false;
@@ -159,7 +159,7 @@ bool CItemData::DelRecordStatic(CMysqlConnection* pDB, OBJID idItem)
 
 uint32_t CItemData::GetExtra(uint32_t nIdx) const
 {
-	CHECKF(nIdx >= 0 && nIdx < MAX_ITEM_EXTRDATA_NUM && nIdx < m_ExtraData.data_size());
+	CHECKF(nIdx >= 0 && nIdx < MAX_ITEM_EXTRDATA_NUM && nIdx < (uint32_t)m_ExtraData.data_size());
 	return m_ExtraData.data(nIdx);
 }
